% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_shiny_app.R
\name{app_server}
\alias{app_server}
\title{Application Server Logic}
\usage{
app_server(input, output, session)
}
\arguments{
\item{input}{Shiny input object}

\item{output}{Shiny output object}

\item{session}{Shiny session object}
}
\description{
Defines the server-side logic for the clustering application
}
\keyword{internal}
