% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_utils.R
\name{apply_standardization}
\alias{apply_standardization}
\title{Apply standardization using pre-computed parameters}
\usage{
apply_standardization(data, centers, scales)
}
\arguments{
\item{data}{A data frame or matrix to standardize}

\item{centers}{Numeric vector of centers (means) from training data}

\item{scales}{Numeric vector of scales (standard deviations) from training data}
}
\value{
Standardized data frame
}
\description{
Apply standardization using pre-computed parameters
}
