% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_base_clusterer.R
\name{BaseClusterer}
\alias{BaseClusterer}
\title{Base Clustering class for Clustering Algorithms}
\description{
R6 abstract class that serves as a base for all clustering algorithm implementations.
it defines common properties and methods that all clustering algorithms should have.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data}}{The dataset to be clustered.}

\item{\code{n_clusters}}{The number of clusters to form.}

\item{\code{fitted}}{Logical flag indicating whether the model has been fitted.}

\item{\code{clusters}}{The resulting clusters after fitting the model.}

\item{\code{standardize}}{Logical flag indicating whether to standardize the data before clustering.}

\item{\code{max_iter}}{The maximum number of iterations for the clustering algorithm.}

\item{\code{tol}}{The tolerance to declare convergence.}

\item{\code{variable_names}}{The names of the variables being clustered.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-BaseClusterer-new}{\code{BaseClusterer$new()}}
\item \href{#method-BaseClusterer-fit}{\code{BaseClusterer$fit()}}
\item \href{#method-BaseClusterer-predict}{\code{BaseClusterer$predict()}}
\item \href{#method-BaseClusterer-print}{\code{BaseClusterer$print()}}
\item \href{#method-BaseClusterer-summary}{\code{BaseClusterer$summary()}}
\item \href{#method-BaseClusterer-get_n_variables}{\code{BaseClusterer$get_n_variables()}}
\item \href{#method-BaseClusterer-get_n_observations}{\code{BaseClusterer$get_n_observations()}}
\item \href{#method-BaseClusterer-get_cluster_sizes}{\code{BaseClusterer$get_cluster_sizes()}}
\item \href{#method-BaseClusterer-get_cluster_members}{\code{BaseClusterer$get_cluster_members()}}
\item \href{#method-BaseClusterer-reset}{\code{BaseClusterer$reset()}}
\item \href{#method-BaseClusterer-get_results}{\code{BaseClusterer$get_results()}}
\item \href{#method-BaseClusterer-save_results}{\code{BaseClusterer$save_results()}}
\item \href{#method-BaseClusterer-get_centers}{\code{BaseClusterer$get_centers()}}
\item \href{#method-BaseClusterer-clone}{\code{BaseClusterer$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BaseClusterer-new"></a>}}
\if{latex}{\out{\hypertarget{method-BaseClusterer-new}{}}}
\subsection{Method \code{new()}}{
Create a new BaseClusterer object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseClusterer$new(
  data,
  n_clusters,
  standardize = FALSE,
  max_iter = 100,
  tol = 1e-04,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{The dataset to be clustered (variables in columns, observations in rows).}

\item{\code{n_clusters}}{The number of clusters to form.}

\item{\code{standardize}}{Whether to standardize the data.}

\item{\code{max_iter}}{Maximum number of iterations.}

\item{\code{tol}}{Tolerance for convergence.}

\item{\code{...}}{Additional parameters for specific clusterer implementations.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{BaseClusterer} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BaseClusterer-fit"></a>}}
\if{latex}{\out{\hypertarget{method-BaseClusterer-fit}{}}}
\subsection{Method \code{fit()}}{
Fit the clustering algorithm to the data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseClusterer$fit()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The fitted clustering model.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BaseClusterer-predict"></a>}}
\if{latex}{\out{\hypertarget{method-BaseClusterer-predict}{}}}
\subsection{Method \code{predict()}}{
Predict the cluster labels for new data points.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseClusterer$predict(new_data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{new_data}}{New data points to predict cluster labels for.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A vector of predicted cluster labels.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BaseClusterer-print"></a>}}
\if{latex}{\out{\hypertarget{method-BaseClusterer-print}{}}}
\subsection{Method \code{print()}}{
model display
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseClusterer$print()}\if{html}{\out{</div>}}
}

\subsection{Details}{
This method should be overridden in derived classes
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BaseClusterer-summary"></a>}}
\if{latex}{\out{\hypertarget{method-BaseClusterer-summary}{}}}
\subsection{Method \code{summary()}}{
summary of the model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseClusterer$summary()}\if{html}{\out{</div>}}
}

\subsection{Details}{
This method should be overridden in derived classes
}

\subsection{Returns}{
A summary of the clustering model.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BaseClusterer-get_n_variables"></a>}}
\if{latex}{\out{\hypertarget{method-BaseClusterer-get_n_variables}{}}}
\subsection{Method \code{get_n_variables()}}{
Get the number of variables in the dataset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseClusterer$get_n_variables()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Integer
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BaseClusterer-get_n_observations"></a>}}
\if{latex}{\out{\hypertarget{method-BaseClusterer-get_n_observations}{}}}
\subsection{Method \code{get_n_observations()}}{
Get the number of observations in the dataset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseClusterer$get_n_observations()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Integer
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BaseClusterer-get_cluster_sizes"></a>}}
\if{latex}{\out{\hypertarget{method-BaseClusterer-get_cluster_sizes}{}}}
\subsection{Method \code{get_cluster_sizes()}}{
Get cluster sizes
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseClusterer$get_cluster_sizes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Named vector of cluster sizes
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BaseClusterer-get_cluster_members"></a>}}
\if{latex}{\out{\hypertarget{method-BaseClusterer-get_cluster_members}{}}}
\subsection{Method \code{get_cluster_members()}}{
Get variables in a specific cluster
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseClusterer$get_cluster_members(cluster_id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cluster_id}}{Integer, cluster number}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Character vector of variable names
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BaseClusterer-reset"></a>}}
\if{latex}{\out{\hypertarget{method-BaseClusterer-reset}{}}}
\subsection{Method \code{reset()}}{
Reset the model to unfitted state
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseClusterer$reset()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BaseClusterer-get_results"></a>}}
\if{latex}{\out{\hypertarget{method-BaseClusterer-get_results}{}}}
\subsection{Method \code{get_results()}}{
Export clustering results as data.frame
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseClusterer$get_results()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Data frame with variable names and cluster assignments
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BaseClusterer-save_results"></a>}}
\if{latex}{\out{\hypertarget{method-BaseClusterer-save_results}{}}}
\subsection{Method \code{save_results()}}{
Export results to CSV file
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseClusterer$save_results(file_path)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{file_path}}{Path to save the CSV file}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BaseClusterer-get_centers"></a>}}
\if{latex}{\out{\hypertarget{method-BaseClusterer-get_centers}{}}}
\subsection{Method \code{get_centers()}}{
Get cluster centers (for compatibility with K selection methods)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseClusterer$get_centers()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Matrix of cluster centers or NULL for base class
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BaseClusterer-clone"></a>}}
\if{latex}{\out{\hypertarget{method-BaseClusterer-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseClusterer$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
