% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_cluster_validator.R
\name{calinski_harabasz_method}
\alias{calinski_harabasz_method}
\title{Calinski-Harabasz Pseudo-F (Adapted for Variable Clustering)}
\usage{
calinski_harabasz_method(
  clusterer_class,
  data,
  k_range = 2:10,
  plot = TRUE,
  ...
)
}
\arguments{
\item{clusterer_class}{R6 class of the clusterer.}

\item{data}{Data frame to be clustered.}

\item{k_range}{Integer vector of K values to test.}

\item{plot}{Logical. Whether to display the plot.}

\item{...}{Additional parameters for the clusterer.}
}
\value{
A list with the following components:
\item{k_values}{Numeric vector of the K values that were tested.}
\item{ch_scores}{Numeric vector of the Pseudo-F statistics
corresponding to each K value. Higher values indicate better clustering.}
\item{suggested_k}{The K value with the highest Pseudo-F score.}
}
\description{
Calculates a Pseudo-F statistic based on the ratio of explained
to unexplained variance, penalized by the number of clusters.
}
\details{
The Calinski-Harabasz index (Pseudo-F) balances clustering quality with
parsimony. It uses the ratio of explained to unexplained variance, penalized
by the number of clusters:

\deqn{CH(k) = \frac{(H - 1)/(k-1)}{(n_{vars} - H)/(n_{vars} - k)}}

where \eqn{H} is the total explained inertia (sum of within-cluster homogeneities)
and \eqn{n_{vars}} is the total number of variables.

Higher values indicate better-defined, more separated clusters. The optimal K
typically corresponds to the maximum Pseudo-F value.
}
\seealso{
\code{\link{elbow_method}}, \code{\link{silhouette_method}},
\code{\link{compare_k_selection_methods}}
}
