% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_cluster_validator.R
\name{cluster-validator}
\alias{cluster-validator}
\title{Cluster Validation and K Selection Methods}
\description{
Functions to help choose the optimal number of clusters using
various statistical methods (Elbow, Silhouette, Homogeneity-based).
}
