% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_utils.R
\name{compare_k_selection_methods}
\alias{compare_k_selection_methods}
\title{Comparison of all K selection methods}
\usage{
compare_k_selection_methods(
  clusterer_class,
  data,
  k_range = 2:10,
  plot = TRUE,
  ...
)
}
\arguments{
\item{clusterer_class}{R6 class of the clusterer}

\item{data}{Dataset to be clustered}

\item{k_range}{Vector of K values to test (default: 2:10)}

\item{plot}{Logical, whether to display plots (default: TRUE)}

\item{...}{Additional parameters for the clusterer}
}
\value{
List with results from all methods
}
\description{
Runs all selection methods and displays a comparative summary.
}
