% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_cluster_validator.R
\name{elbow_method}
\alias{elbow_method}
\title{Elbow Method for Optimal K Selection}
\usage{
elbow_method(clusterer_class, data, k_range = 2:10, plot = TRUE, ...)
}
\arguments{
\item{clusterer_class}{R6 class of the clusterer (e.g., \code{KMeansClusterer}).}

\item{data}{Data frame or matrix to be clustered.}

\item{k_range}{Integer vector of K values to test (default: \code{2:10}).
Must contain at least 2 values.}

\item{plot}{Logical. Whether to display the elbow plot (default: \code{TRUE}).}

\item{...}{Additional parameters passed to the clusterer constructor.}
}
\value{
A list with the following components:
\item{k_values}{Numeric vector of the K values that were tested.}
\item{inertias}{Numeric vector of the quality criterion (lower is better)
corresponding to each K value. For KMeansClusterer, this is (1 - homogeneity).}
\item{suggested_k}{The K value identified by the elbow method as the
potential optimum.}
\item{distances}{Numeric vector of perpendicular distances of each K-point
to the start-end line, used to find the elbow.}
}
\description{
Calculates a quality criterion for different values of K and identifies
the "elbow" of the curve which suggests the optimal number of clusters.
For variable clustering (KMeansClusterer), uses (1 - homogeneity) as criterion.
}
\seealso{
\code{\link{silhouette_method}}, \code{\link{calinski_harabasz_method}},
\code{\link{compare_k_selection_methods}}
}
