% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_kmeans.R
\name{kmeans_simple}
\alias{kmeans_simple}
\title{simple kmeans clustering}
\usage{
kmeans_simple(x, centers = 3)
}
\arguments{
\item{x}{matrix or data.frame}

\item{centers}{number of centers}
}
\value{
vector of cluster assignments
}
\description{
simple kmeans clustering
}
