% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_mca_hclust_cluster.R
\name{MCA_HClusterer}
\alias{MCA_HClusterer}
\title{Clustering using MCA followed by HAC (Hierarchical Agglomerative Clustering) for qualitative data}
\description{
R6 class that implements clustering by first performing Multiple Correspondence Analysis (MCA)
on qualitative data, followed by Hierarchical Agglomerative Clustering (HAC) on the MCA results.
}
\section{Super class}{
\code{\link[M2RClust:BaseClusterer]{M2RClust::BaseClusterer}} -> \code{MCA_HClusterer}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{mca_result}}{The result of the MCA analysis.}

\item{\code{hclust_result}}{The result of the hierarchical clustering.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MCA_HClusterer-new}{\code{MCA_HClusterer$new()}}
\item \href{#method-MCA_HClusterer-fit}{\code{MCA_HClusterer$fit()}}
\item \href{#method-MCA_HClusterer-predict}{\code{MCA_HClusterer$predict()}}
\item \href{#method-MCA_HClusterer-print}{\code{MCA_HClusterer$print()}}
\item \href{#method-MCA_HClusterer-summary}{\code{MCA_HClusterer$summary()}}
\item \href{#method-MCA_HClusterer-plot_dendrogram}{\code{MCA_HClusterer$plot_dendrogram()}}
\item \href{#method-MCA_HClusterer-clone}{\code{MCA_HClusterer$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="M2RClust" data-topic="BaseClusterer" data-id="get_centers"><a href='../../M2RClust/html/BaseClusterer.html#method-BaseClusterer-get_centers'><code>M2RClust::BaseClusterer$get_centers()</code></a></span></li>
<li><span class="pkg-link" data-pkg="M2RClust" data-topic="BaseClusterer" data-id="get_cluster_members"><a href='../../M2RClust/html/BaseClusterer.html#method-BaseClusterer-get_cluster_members'><code>M2RClust::BaseClusterer$get_cluster_members()</code></a></span></li>
<li><span class="pkg-link" data-pkg="M2RClust" data-topic="BaseClusterer" data-id="get_cluster_sizes"><a href='../../M2RClust/html/BaseClusterer.html#method-BaseClusterer-get_cluster_sizes'><code>M2RClust::BaseClusterer$get_cluster_sizes()</code></a></span></li>
<li><span class="pkg-link" data-pkg="M2RClust" data-topic="BaseClusterer" data-id="get_n_observations"><a href='../../M2RClust/html/BaseClusterer.html#method-BaseClusterer-get_n_observations'><code>M2RClust::BaseClusterer$get_n_observations()</code></a></span></li>
<li><span class="pkg-link" data-pkg="M2RClust" data-topic="BaseClusterer" data-id="get_n_variables"><a href='../../M2RClust/html/BaseClusterer.html#method-BaseClusterer-get_n_variables'><code>M2RClust::BaseClusterer$get_n_variables()</code></a></span></li>
<li><span class="pkg-link" data-pkg="M2RClust" data-topic="BaseClusterer" data-id="get_results"><a href='../../M2RClust/html/BaseClusterer.html#method-BaseClusterer-get_results'><code>M2RClust::BaseClusterer$get_results()</code></a></span></li>
<li><span class="pkg-link" data-pkg="M2RClust" data-topic="BaseClusterer" data-id="reset"><a href='../../M2RClust/html/BaseClusterer.html#method-BaseClusterer-reset'><code>M2RClust::BaseClusterer$reset()</code></a></span></li>
<li><span class="pkg-link" data-pkg="M2RClust" data-topic="BaseClusterer" data-id="save_results"><a href='../../M2RClust/html/BaseClusterer.html#method-BaseClusterer-save_results'><code>M2RClust::BaseClusterer$save_results()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MCA_HClusterer-new"></a>}}
\if{latex}{\out{\hypertarget{method-MCA_HClusterer-new}{}}}
\subsection{Method \code{new()}}{
Create a new MCA_HClusterer object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MCA_HClusterer$new(data, n_clusters)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{The dataset to be clustered.}

\item{\code{n_clusters}}{The number of clusters to form.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{MCA_HClusterer} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MCA_HClusterer-fit"></a>}}
\if{latex}{\out{\hypertarget{method-MCA_HClusterer-fit}{}}}
\subsection{Method \code{fit()}}{
Fit the MCA followed by HAC clustering algorithm to the data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MCA_HClusterer$fit()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The fitted MCA + HAC clustering model.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MCA_HClusterer-predict"></a>}}
\if{latex}{\out{\hypertarget{method-MCA_HClusterer-predict}{}}}
\subsection{Method \code{predict()}}{
Predict the cluster labels for new data points using the fitted MCA + HAC model.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MCA_HClusterer$predict(new_data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{new_data}}{New data points to predict cluster labels for.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A vector of predicted cluster labels.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MCA_HClusterer-print"></a>}}
\if{latex}{\out{\hypertarget{method-MCA_HClusterer-print}{}}}
\subsection{Method \code{print()}}{
model display
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MCA_HClusterer$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MCA_HClusterer-summary"></a>}}
\if{latex}{\out{\hypertarget{method-MCA_HClusterer-summary}{}}}
\subsection{Method \code{summary()}}{
summary of the MCA + HAC model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MCA_HClusterer$summary()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A summary of the MCA + HAC clustering model.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MCA_HClusterer-plot_dendrogram"></a>}}
\if{latex}{\out{\hypertarget{method-MCA_HClusterer-plot_dendrogram}{}}}
\subsection{Method \code{plot_dendrogram()}}{
plot dendrogram of the hierarchical clustering
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MCA_HClusterer$plot_dendrogram()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A dendrogram plot.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MCA_HClusterer-clone"></a>}}
\if{latex}{\out{\hypertarget{method-MCA_HClusterer-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MCA_HClusterer$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
