% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_mixed_clusterer.R
\name{MixedClusterer}
\alias{MixedClusterer}
\title{Clustering class for Mixed Data (method still to be defined)}
\description{
R6 class that implements clustering for mixed data (both qualitative and quantitative variables).
}
\section{Super class}{
\code{\link[M2RClust:BaseClusterer]{M2RClust::BaseClusterer}} -> \code{MixedClusterer}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{method}}{The clustering method to be used for mixed data.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MixedClusterer-new}{\code{MixedClusterer$new()}}
\item \href{#method-MixedClusterer-fit}{\code{MixedClusterer$fit()}}
\item \href{#method-MixedClusterer-predict}{\code{MixedClusterer$predict()}}
\item \href{#method-MixedClusterer-print}{\code{MixedClusterer$print()}}
\item \href{#method-MixedClusterer-summary}{\code{MixedClusterer$summary()}}
\item \href{#method-MixedClusterer-clone}{\code{MixedClusterer$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="M2RClust" data-topic="BaseClusterer" data-id="get_centers"><a href='../../M2RClust/html/BaseClusterer.html#method-BaseClusterer-get_centers'><code>M2RClust::BaseClusterer$get_centers()</code></a></span></li>
<li><span class="pkg-link" data-pkg="M2RClust" data-topic="BaseClusterer" data-id="get_cluster_members"><a href='../../M2RClust/html/BaseClusterer.html#method-BaseClusterer-get_cluster_members'><code>M2RClust::BaseClusterer$get_cluster_members()</code></a></span></li>
<li><span class="pkg-link" data-pkg="M2RClust" data-topic="BaseClusterer" data-id="get_cluster_sizes"><a href='../../M2RClust/html/BaseClusterer.html#method-BaseClusterer-get_cluster_sizes'><code>M2RClust::BaseClusterer$get_cluster_sizes()</code></a></span></li>
<li><span class="pkg-link" data-pkg="M2RClust" data-topic="BaseClusterer" data-id="get_n_observations"><a href='../../M2RClust/html/BaseClusterer.html#method-BaseClusterer-get_n_observations'><code>M2RClust::BaseClusterer$get_n_observations()</code></a></span></li>
<li><span class="pkg-link" data-pkg="M2RClust" data-topic="BaseClusterer" data-id="get_n_variables"><a href='../../M2RClust/html/BaseClusterer.html#method-BaseClusterer-get_n_variables'><code>M2RClust::BaseClusterer$get_n_variables()</code></a></span></li>
<li><span class="pkg-link" data-pkg="M2RClust" data-topic="BaseClusterer" data-id="get_results"><a href='../../M2RClust/html/BaseClusterer.html#method-BaseClusterer-get_results'><code>M2RClust::BaseClusterer$get_results()</code></a></span></li>
<li><span class="pkg-link" data-pkg="M2RClust" data-topic="BaseClusterer" data-id="reset"><a href='../../M2RClust/html/BaseClusterer.html#method-BaseClusterer-reset'><code>M2RClust::BaseClusterer$reset()</code></a></span></li>
<li><span class="pkg-link" data-pkg="M2RClust" data-topic="BaseClusterer" data-id="save_results"><a href='../../M2RClust/html/BaseClusterer.html#method-BaseClusterer-save_results'><code>M2RClust::BaseClusterer$save_results()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MixedClusterer-new"></a>}}
\if{latex}{\out{\hypertarget{method-MixedClusterer-new}{}}}
\subsection{Method \code{new()}}{
Create a new MixedClusterer object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MixedClusterer$new(data, n_clusters)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{The dataset to be clustered.}

\item{\code{n_clusters}}{The number of clusters to form.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{MixedClusterer} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MixedClusterer-fit"></a>}}
\if{latex}{\out{\hypertarget{method-MixedClusterer-fit}{}}}
\subsection{Method \code{fit()}}{
Fit the mixed data clustering algorithm to the data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MixedClusterer$fit()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The fitted mixed data clustering model.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MixedClusterer-predict"></a>}}
\if{latex}{\out{\hypertarget{method-MixedClusterer-predict}{}}}
\subsection{Method \code{predict()}}{
Predict the cluster labels for new data points using the fitted mixed data model.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MixedClusterer$predict(new_data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{new_data}}{New data points to predict cluster labels for.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A vector of predicted cluster labels.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MixedClusterer-print"></a>}}
\if{latex}{\out{\hypertarget{method-MixedClusterer-print}{}}}
\subsection{Method \code{print()}}{
model display
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MixedClusterer$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MixedClusterer-summary"></a>}}
\if{latex}{\out{\hypertarget{method-MixedClusterer-summary}{}}}
\subsection{Method \code{summary()}}{
summary of the Mixed data model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MixedClusterer$summary()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A summary of the Mixed data clustering model.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MixedClusterer-clone"></a>}}
\if{latex}{\out{\hypertarget{method-MixedClusterer-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MixedClusterer$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
