% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_visualization.R
\name{plot_clustering_with_supp}
\alias{plot_clustering_with_supp}
\title{Plot clustering results with supplementary variables}
\usage{
plot_clustering_with_supp(
  clusterer,
  main = NULL,
  show_centers = TRUE,
  show_labels = TRUE
)
}
\arguments{
\item{clusterer}{A fitted clustering object with prediction made}

\item{main}{Title of the plot}

\item{show_centers}{Logical, whether to show cluster centers}

\item{show_labels}{Logical, whether to show variable labels}
}
\description{
Plot clustering results with supplementary variables
}
