% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_visualization.R
\name{plot_correlation_heatmap}
\alias{plot_correlation_heatmap}
\title{Plot heatmap of correlations by cluster}
\usage{
plot_correlation_heatmap(
  clusterer,
  main = "Correlation Heatmap by Cluster",
  reorder = TRUE
)
}
\arguments{
\item{clusterer}{A fitted clustering object}

\item{main}{Title of the plot}

\item{reorder}{Logical, whether to reorder variables by cluster}
}
\description{
Plot heatmap of correlations by cluster
}
