% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_visualization.R
\name{plot_dendrogram}
\alias{plot_dendrogram}
\title{Plot dendrogram for hierarchical clustering}
\usage{
plot_dendrogram(clusterer, main = "Dendrogram", ...)
}
\arguments{
\item{clusterer}{A fitted hierarchical clustering object}

\item{main}{Title of the plot}

\item{...}{Additional parameters passed to plot()}
}
\description{
Plot dendrogram for hierarchical clustering
}
