% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_visualization.R
\name{plot_scree_by_cluster}
\alias{plot_scree_by_cluster}
\title{Plot scree plot of explained variance by cluster}
\usage{
plot_scree_by_cluster(clusterer, main = "Scree Plot by Cluster")
}
\arguments{
\item{clusterer}{A fitted clustering object}

\item{main}{Title of the plot}
}
\description{
Plot scree plot of explained variance by cluster
}
