% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_visualization.R
\name{plot_variable_contributions}
\alias{plot_variable_contributions}
\title{Plot variable contributions to clusters}
\usage{
plot_variable_contributions(
  clusterer,
  cluster_id = NULL,
  top_n = 10,
  main = "Variable Contributions"
)
}
\arguments{
\item{clusterer}{A fitted clustering object}

\item{cluster_id}{Cluster number to visualize (NULL for all clusters)}

\item{top_n}{Number of top contributing variables to show}

\item{main}{Title of the plot}
}
\description{
Plot variable contributions to clusters
}
