% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_utils.R
\name{standardize_data}
\alias{standardize_data}
\title{Utility Functions for Clustering}
\usage{
standardize_data(data, center = TRUE, scale = TRUE)
}
\arguments{
\item{data}{A data frame or matrix to standardize}

\item{center}{Logical or numeric vector of centers. If TRUE, centering is done by subtracting the column means.}

\item{scale}{Logical or numeric vector of scales. If TRUE, scaling is done by dividing by the standard deviations.}
}
\value{
A list containing the scaled data, the centers, and the scales
}
\description{
This script contains utility functions used by all the implemented clustering algorithms.
}
\keyword{(center}
\keyword{Standardize}
\keyword{and}
\keyword{data}
\keyword{internal}
\keyword{scale)}
