% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_utils.R
\name{elbow_method}
\alias{elbow_method}
\title{Elbow Method}
\usage{
elbow_method(clusterer_class, data, k_range = 2:10, plot = TRUE, ...)
}
\arguments{
\item{clusterer_class}{R6 class of the clusterer (e.g., KMeansClusterer)}

\item{data}{Dataset to be clustered}

\item{k_range}{Vector of K values to test (default: 2:10)}

\item{plot}{Logical, whether to display the plot (default: TRUE)}

\item{...}{Additional parameters for the clusterer}
}
\value{
List with inertias, k_values, and suggested_k
}
\description{
Calculates the within-cluster inertia for different values of K and identifies
the "elbow" of the curve which suggests the optimal number of clusters.
}
