% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_visualization.R
\name{plot_cluster_profiles_lines}
\alias{plot_cluster_profiles_lines}
\title{Fallback line plot for cluster profiles}
\usage{
plot_cluster_profiles_lines(clusterer, main, standardize)
}
\description{
Fallback line plot for cluster profiles
}
\keyword{internal}
