% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_visualization.R
\name{plot_cluster_quality}
\alias{plot_cluster_quality}
\title{Plot cluster sizes and quality metrics}
\usage{
plot_cluster_quality(clusterer, main = "Cluster Quality Metrics")
}
\arguments{
\item{clusterer}{A fitted clustering object}

\item{main}{Title of the plot}
}
\description{
Plot cluster sizes and quality metrics
}
