% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_visualization.R
\name{plot_clustering_2d}
\alias{plot_clustering_2d}
\title{Plot clustering results in 2D PCA space}
\usage{
plot_clustering_2d(
  clusterer,
  main = NULL,
  show_centers = TRUE,
  show_labels = TRUE
)
}
\arguments{
\item{clusterer}{A fitted clustering object (e.g., KMeansClusterer)}

\item{main}{Title of the plot}

\item{show_centers}{Logical, whether to show cluster centers}

\item{show_labels}{Logical, whether to show variable labels}
}
\description{
Plot clustering results in 2D PCA space for fitted model
}
