% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_visualization.R
\name{plot_radar_chart}
\alias{plot_radar_chart}
\title{Plot radar chart comparing clusters}
\usage{
plot_radar_chart(
  clusterer,
  main = "Cluster Profiles (Radar Chart)",
  standardize = TRUE
)
}
\arguments{
\item{clusterer}{A fitted clustering object}

\item{main}{Title of the plot}

\item{standardize}{Logical, whether to standardize variables before plotting}
}
\description{
Plot radar chart comparing clusters
}
