% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_utils.R
\name{silhouette_method}
\alias{silhouette_method}
\title{Silhouette Method}
\usage{
silhouette_method(clusterer_class, data, k_range = 2:10, plot = TRUE, ...)
}
\arguments{
\item{clusterer_class}{R6 class of the clusterer}

\item{data}{Dataset to be clustered}

\item{k_range}{Vector of K values to test (default: 2:10)}

\item{plot}{Logical, whether to display the plot (default: TRUE)}

\item{...}{Additional parameters for the clusterer}
}
\value{
List with silhouette_scores, k_values, and suggested_k
}
\description{
Calculates the average silhouette coefficient for different values of K.
The optimal K is the one that maximizes this coefficient.
}
